requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ec==0.2.5
h11==0.14.0
icrawler==0.6.10
idna==3.10
imageio==2.37.0
joblib==1.4.2
lazy_loader==0.4
lxml==5.3.1
networkx==3.4.2
numpy==2.2.4
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
pycparser==2.22
PySocks==1.7.1
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.31.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
threadpoolctl==3.6.0
tifffile==2025.3.30
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.0
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
sphinx==7.3.7
sphinx-rtd-theme==2.0.0
myst-parser==2.0.0